# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hospital Management System API.",
        "title": "HMS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/patients": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Receptionists and Doctors can list all registered patients.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "List all registered patients",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Limit (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset (default: 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Patient"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid pagination parameters",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    }
                }
            }
        },
        "/patients/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Doctors can register a new patient.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Register a new patient",
                "parameters": [
                    {
                        "description": "Patient Registration Data",
                        "name": "patientRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PatientCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Patient"
                        }
                    },
                    "400": {
                        "description": "Validation error or invalid input",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    }
                }
            }
        },
        "/patients/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Receptionists and Doctors can get details of a specific patient.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Get details of a specific patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Patient"
                        }
                    },
                    "400": {
                        "description": "Invalid patient ID format",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "404": {
                        "description": "Patient not found",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Only doctors can delete patient records",
                "tags": [
                    "Patients"
                ],
                "summary": "Delete a patient record",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Patient ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Patient record deleted successfully"
                    },
                    "400": {
                        "description": "Validation error, invalid input, or invalid patient ID",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden (e.g., if trying to update restricted fields)",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "404": {
                        "description": "Patient not found",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Receptionists can update most patient details. Doctors can update patient details, especially medical history.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Update patient details",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Patient ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Patient Update Data (fields to update)",
                        "name": "patientRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PatientUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Patient"
                        }
                    },
                    "400": {
                        "description": "Validation error, invalid input, or invalid patient ID",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden (e.g., if trying to update restricted fields)",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "404": {
                        "description": "Patient not found",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.APIError": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Can be map[string]string for validation errors, or a simple string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.Gender": {
            "type": "string",
            "enum": [
                "male",
                "female",
                "other"
            ],
            "x-enum-varnames": [
                "GenderMale",
                "GenderFemale",
                "GenderOther"
            ]
        },
...
        "model.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "The actual list of items"
                },
                "limit": {
                    "description": "The limit used for this page",
                    "type": "integer"
                },
                "offset": {
                    "description": "The offset used for this page",
                    "type": "integer"
                },
                "page": {
                    "description": "Current page number (calculated: offset/limit + 1)",
                    "type": "integer"
                },
                "total": {
                    "description": "Total number of items available across all pages",
                    "type": "integer"
                },
                "total_pages": {
                    "description": "Total pages (calculated: ceil(total/limit))",
                    "type": "integer"
                }
            }
        },
        "model.Patient": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "contact_email": {
                    "type": "string"
                },
                "contact_phone": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date_of_birth": {
                    "description": "Recommended to store as full timestamp, format for display",
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/model.Gender"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "medical_history": {
                    "type": "string"
                },
                "registered_by_user_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.PatientCreateRequest": {
            "type": "object",
            "required": [
                "date_of_birth",
                "first_name",
                "last_name"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 500
                },
                "contact_email": {
                    "type": "string",
                    "maxLength": 255
                },
                "contact_phone": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "gender": {
                    "enum": [
                        "male",
                        "female",
                        "other"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Gender"
                        }
                    ]
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "medical_history": {
                    "type": "string"
                }
            }
        },
        "model.PatientUpdateRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 500
                },
                "contact_email": {
                    "type": "string",
                    "maxLength": 255
                },
                "contact_phone": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "gender": {
                    "enum": [
                        "male",
                        "female",
                        "other",
                        "prefer_not_to_say"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Gender"
                        }
                    ]
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "medical_history": {
                    "description": "Doctors can update this",
                    "type": "string"
                }
            }
        },
...
    }
}